[issue-181] Implement Flink TableFactory   * fixed documentation issues   * addressed PR review comments
